Move the unit test to com.google.cloud.firestore
